Staging Commit:
WIIT-238	History Tracking - Maintain Device Mapping history for auditing purpose
WIIT-241	Maintain the history when Device mapped to new or old cart
WIIT-251	Option to remove the mapped simcard from the device with conformation message.
WIIT-253	Messaging - Option to notify the driver to pick the specific cart in the specific location
WIIT-255	UI Screen to sent a message from Store and Company
WIIT-256	UI Screen to display the message from the system to the driver
WIIT-257	Provide option to sent a static reply to the system from the driver
WIIT-258	Display the list of messages already send, In Wiitrak, Prime Company, Corporate, Store logins
WIIT-259	Display the Notification and the Message in the index page at the time of login.